Made csv.hpp/string_view.hpp first sections of single header file
